#4153 Move letter model and repo
#4153 Serial Letter Feature
